# Conflicts:
#	features/transaction/cart/src/main/java/com/tokopedia/cartrevamp/view/uimodel/CartItemHolderData.kt

+++ features/discovery/search/src/main/java/com/tokopedia/search/result/product/pagination/PaginationImpl.kt
@@ -9,7 +9,7 @@ import javax.inject.Inject
 class PaginationImpl @Inject constructor(): Pagination {
     override var startFrom: Int = 0
 
-    override var totalData: Int = 0
+    override var totalData: Long = 0L
 
     override fun clearData() {
         startFrom = 0
